Basic Crossover and Mutation Working
Crossover and mutation of enemies is working in this version.

Mutation of attributes within certain ranges are between 10% less than
min and 10% greater than max. May need tweaking for mutation to work as
desired.

Enemies in a wave are spawned (almost) simultaneously rather than
waiting for a set time between enemies. This is because
CrossoverAndMutate is not a coroutine.

Other changes include:
Camera spawn parameters now contained in GameController script.
Some unnecessary methods removed.