Basic Fitness Evaluation (not used in evolution yet)
Fitness scores are now calcualted based on lifetime of the enemy. These
scores are not used yet in the process of choosing which enemies to
crossover.

Future work will use fitness values to decide on which enemies will be
used in reproduction and how often. Possible strategies include: Only
crossing over the top 50% of enemies / cross-over each enemy with the
fittest individual only etc.