Accuracy and On Target Shots Fitness Functions (+more)
There are 2 new fitness functions:
Accuracy - The percentage of shots an enemy has within a certain
distance of the player.
Number of On Target Shots - The number of shots an enemy has within a
certain distance of the player.

Fitness values are now calculated seperately to having the fitness
property of each individual set to avoid duplicated code and confusion.

Object pooling is now done by a seperate 'Object Pooler' game object
with children for each item that needs to be pooled including Enemy
Ships, Bolts, and Enemy Bolts.