Added Player Health
The player ship now has a health attribute (set on the Player game
object). The player loses 1 health every time it is hit, but won't be
destroyed until their health reaches 0. Health is displayed in the
bottom right corner of the screen.

The background has been made smaller and tiled in order to improve its
resolution.