Destroy Objects Outside Screen
The destroy boundary is now sized based off of screen resolution. As
such, and resolution is now permitted, and the boundary will be sized
accordingly. Previously, only some resolutions would work, as too small
would cause enemies going of the screen to never be destroyed, but too
large would cause the player's ship to be destroyed if travelling near
the edge of the screen.

Also ready for first beta release build.